Add files via upload
This Databricks notebook fetches real-time data from the Transport for London (TFL) Open API to monitor disruptions on the TFL underground network. It processes the data to extract line names, status severity descriptions, and reasons for disruptions, if any. The notebook then analyzes the status of each tube line, categorizing them as "Good Service," "Minor Delays," or "Severe Delays." Finally, it displays the results in a tabular format with timestamps. This provides users with up-to-date information to make informed travel decisions.